completed with patient record and block history pages

+++ HalloDoc/HalloDoc.Models/ViewModels/PhysicianAccountViewModel.cs
@@ -103,7 +103,7 @@ namespace HalloDoc.ViewModels
 
         public string? UserName { get; set; }
 
-        public short? Status { get; set; }
+        public int? Status { get; set; }
 
         public int? role_id { get; set; }
